feat: refactor islands to use nanostores for state management
