+++ client/src/components/SoftwareFilters.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useRef, useState } from "react";
 import { Combobox } from "@/components/Combobox.tsx";
-import { Input } from "@/components/ui/input";
 import { useTranslations } from "../i18n/utils";
+import { selectedOrganisations as selectedOrganisations$ } from "@/stores/filters.ts";
 
 export type Locale = "en" | "de" | "fr" | "it";
 
@@ -23,17 +23,11 @@ export type Department = {
 type Props = {
   lang: Locale;
   organisations: Department[];
-  onSelectedOrganisationsChange: (values: string[]) => void;
-  nameQuery: string;
-  onNameQueryChange: (value: string) => void;
 };
 
 export function SoftwareFilters({
   lang,
   organisations,
-  onSelectedOrganisationsChange,
-  nameQuery,
-  onNameQueryChange,
 }: Props) {
   const [query] = useState("");
   const [_open, setOpen] = useState(false);
@@ -80,39 +74,15 @@ export function SoftwareFilters({
   }, []);
 
   return (
-    <div className="container">
-      <div className="search">
-        <div className="search__filters filters--are-open">
-          <div className="grid grid--responsive-cols-2 gap--responsive">
-            <div
-              className="form__group__select"
-              ref={dropdownRef}
-              style={{ position: "relative" }}
-            >
-              <label className="text--base" htmlFor="organization-filter">
-                {t("index.filterByOrganisation")}
-              </label>
-              <Combobox
-                groups={groupedOptions}
-                lang={lang}
-                onChange={(values) => onSelectedOrganisationsChange(values)}
-              />
-            </div>
-
-            <div className="form__group__input">
-              <label className="text--base" htmlFor="software-name-filter">
-                {t("index.filterByName")}
-              </label>
-              <Input
-                id="software-name-filter"
-                value={nameQuery}
-                onChange={(e) => onNameQueryChange(e.target.value)}
-                placeholder={t("index.filterByName.placeholder")}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+    <>
+      <label className="text--base" htmlFor="organization-filter">
+        {t("index.filterByOrganisation")}
+      </label>
+      <Combobox
+        groups={groupedOptions}
+        lang={lang}
+        onChange={(values) => selectedOrganisations$.set(values)}
+      />
+    </>
   );
 }
